Let non-admin users remove themselves from orgs and projects

+++ admin/server/projects.go
@@ -711,17 +711,20 @@ func (s *Server) RemoveProjectMember(ctx context.Context, req *adminv1.RemovePro
 		return nil, status.Error(codes.InvalidArgument, err.Error())
 	}
 
-	claims := auth.GetClaims(ctx)
-	if !claims.ProjectPermissions(ctx, proj.OrganizationID, proj.ID).ManageProjectMembers {
-		return nil, status.Error(codes.PermissionDenied, "not allowed to remove project members")
-	}
-
 	user, err := s.admin.DB.FindUserByEmail(ctx, req.Email)
 	if err != nil {
 		if !errors.Is(err, database.ErrNotFound) {
 			return nil, status.Error(codes.Internal, err.Error())
 		}
-		// check if there is a pending invite
+
+		// Only admins can remove pending invites.
+		// NOTE: If we change invites to accept/decline (instead of auto-accept on signup), we need to revisit this.
+		claims := auth.GetClaims(ctx)
+		if !claims.ProjectPermissions(ctx, proj.OrganizationID, proj.ID).ManageProjectMembers {
+			return nil, status.Error(codes.PermissionDenied, "not allowed to remove project members")
+		}
+
+		// Check if there is a pending invite
 		invite, err := s.admin.DB.FindProjectInvite(ctx, proj.ID, req.Email)
 		if err != nil {
 			if errors.Is(err, database.ErrNotFound) {
@@ -729,11 +732,21 @@ func (s *Server) RemoveProjectMember(ctx context.Context, req *adminv1.RemovePro
 			}
 			return nil, status.Error(codes.Internal, err.Error())
 		}
+
 		err = s.admin.DB.DeleteProjectInvite(ctx, invite.ID)
 		if err != nil {
 			return nil, status.Error(codes.Internal, err.Error())
 		}
 		return &adminv1.RemoveProjectMemberResponse{}, nil
+	}
+
+	// The caller must either have ManageProjectMembers permission or be the user being removed.
+	claims := auth.GetClaims(ctx)
+	allow := false
+	allow = allow || claims.ProjectPermissions(ctx, proj.OrganizationID, proj.ID).ManageProjectMembers
+	allow = allow || (claims.OwnerType() == auth.OwnerTypeUser && claims.OwnerID() == user.ID)
+	if !allow {
+		return nil, status.Error(codes.PermissionDenied, "not allowed to remove project members")
 	}
 
 	err = s.admin.DB.DeleteProjectMemberUser(ctx, proj.ID, user.ID)
